+++ sdk/core/azure-core/samples/test_example_sansio.py
@@ -167,16 +167,3 @@ def example_proxy_policy():
     # You can also configure proxies by setting the environment variables
     # HTTP_PROXY and HTTPS_PROXY.
     # [END proxy_policy]
-
-def example_on_exception():
-    policy = SansIOHTTPPolicy()
-    request = HttpRequest("GET", "https://bing.com")
-    # [START on_exception]
-    try:
-        response = policy.on_request(request)
-    except Exception:
-        policy.on_exception(request)
-
-    # or use
-    exc_type, exc_value, exc_traceback = sys.exc_info()
-    # [END on_exception]